Split config from datastore into own module
